빌드 버전 오류 해결
8.2.2 = AGP 8.11버전

+++ android-pdf-writer/src/main/java/com/hangyeolee/androidpdfwriter/components/PDFGridLayout.java
@@ -152,7 +152,7 @@ public class PDFGridLayout extends PDFLayout{
                 if(i == 0) {
                     margin.top += (gap);
                     updateHeight(gap);
-                    int d = 0;
+                    float d = 0;
                     if (parent != null)
                         d += parent.measureY + parent.border.size.top + parent.padding.top;
                     measureY = relativeY + margin.top + d;
